requirements-dev.txt
-e .
bumpversion==0.5.3
djangorestframework==3.6.2
flake8==3.3.0
isort==4.2.5
mixer==5.6.6
mkdocs==0.16.3
mock==2.0.0
pytest-cov==2.4.0
pytest-django==3.1.2
pytest==3.0.7
